% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/samir.R
\docType{package}
\name{samir-package}
\alias{samir-package}
\title{Functions for shuffle and dice rolling}
\description{
Miscelaneous functions for testing code in class.
}
\details{
You shuffle cards or roll dice
}
\section{My Custom Section}{
 Add additional information here.
}
\keyword{package}

