% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle function}
\usage{
shuffle(cards)
}
\arguments{
\item{cards}{are numbers, color and suit.}
}
\value{
This function simulates shuffling a deck of cards
}
\description{
Shuffle function
}
\examples{
site <- "https://gist.githubusercontent.com/garrettgman/9629323/raw/ee5dfc039fd581cb467cc69c226ea2524913c3d8/deck.csv"
deck2 <- readr::read_csv(site)
shuffle(deck2)
deck2 <- shuffle(cards = deck)
deck2[1:5, ]
}

